+++ wasm-crypto-init/wasm.ts
@@ -1,11 +1,11 @@
 // Copyright 2019-2022 @polkadot/wasm-crypto-init authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { InitFn } from 'https://deno.land/x/polkadot@0.0.6/wasm-bridge/types.ts';
+import type { InitFn } from 'https://deno.land/x/polkadot/wasm-bridge/types.ts';
 import type { WasmCryptoInstance } from './types.ts';
 
-import { createWasmFn } from 'https://deno.land/x/polkadot@0.0.6/wasm-bridge/mod.ts';
-import { wasmBytes } from 'https://deno.land/x/polkadot@0.0.6/wasm-crypto-wasm/mod.ts';
+import { createWasmFn } from 'https://deno.land/x/polkadot/wasm-bridge/mod.ts';
+import { wasmBytes } from 'https://deno.land/x/polkadot/wasm-crypto-wasm/mod.ts';
 
 export { packageInfo } from './packageInfo.ts';